\name{homework222-package}
\alias{homework222-package}
\alias{homework222}
\docType{package}
\title{
\packageTitle{homework222}
}
\description{
\packageDescription{homework222}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{homework222}
\packageIndices{homework222}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{homework222}

Maintainer: \packageMaintainer{homework222}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
